AirDrop visual fix
