+++ Content.Shared/Backmen/AirDrop/SharedAirDropSystem.cs
@@ -13,65 +13,17 @@ namespace Content.Shared.Backmen.AirDrop;
 
 public abstract class SharedAirDropSystem : EntitySystem
 {
-    [Dependency] private readonly UseDelaySystem _delay = default!;
+    [Dependency] protected readonly UseDelaySystem Delay = default!;
     [Dependency] private readonly INetManager _net = default!;
-    [Dependency] private readonly EntityTableSystem _entityTable = default!;
-    [Dependency] private readonly IPrototypeManager _prototypeManager = default!;
-    [Dependency] private readonly SharedTransformSystem _transform = default!;
-    [Dependency] private readonly SharedContainerSystem _container = default!;
-    [Dependency] private readonly SharedEntityStorageSystem _storage = default!;
+    [Dependency] protected readonly IPrototypeManager PrototypeManager = default!;
+    [Dependency] protected readonly SharedTransformSystem _transform = default!;
 
     public override void Initialize()
     {
         base.Initialize();
 
-        SubscribeLocalEvent<AirDropItemComponent, AfterInteractEvent>(OnAfterInteract);
         SubscribeLocalEvent<AirDropComponent, MapInitEvent>(OnInitialStartup);
-        SubscribeLocalEvent<AirDropComponent, AirDropSpawnEvent>(OnAirDropSpawn);
         SubscribeLocalEvent<AirDropComponent, AirDropTargetSpawnEvent>(OnAirDropTargetSpawn);
-        SubscribeLocalEvent<AirDropItemComponent, MapInitEvent>(OnMapInit);
-        SubscribeLocalEvent<AirDropItemSpawnEvent>(OnSpawnLoot);
-        SubscribeNetworkEvent<AirDropStartEvent>(OnStartAirDrop);
-    }
-
-    private void OnStartAirDrop(AirDropStartEvent ev)
-    {
-        if (_net.IsServer || !TryGetEntity(ev.Uid, out var supplyPod))
-            return;
-        StartAirDrop(supplyPod.Value, ev.Pos);
-    }
-
-    private void OnSpawnLoot(AirDropItemSpawnEvent ev)
-    {
-        if (ev.Handled)
-            return;
-
-        var supplyTransform = Transform(ev.SupplyPod);
-        var hasContainer =
-            _container.TryGetContainer(ev.SupplyPod, SharedEntityStorageSystem.ContainerName, out var container);
-
-        var openDoorSupplyPod = !hasContainer;
-
-        foreach (var item in _entityTable
-                     .GetSpawns(ev.DropTable))
-        {
-            var entItem = Spawn(item);
-            if (!hasContainer)
-            {
-                _transform.DropNextTo(entItem, ev.SupplyPod);
-                continue;
-            }
-
-            if (!_container.InsertOrDrop(entItem, container!, supplyTransform))
-                openDoorSupplyPod = true;
-        }
-
-        if (openDoorSupplyPod || ev.ForceOpenSupplyDrop)
-        {
-            _storage.OpenStorage(ev.SupplyPod);
-        }
-
-        ev.Handled = true;
     }
 
     private void OnInitialStartup(Entity<AirDropComponent> ent, ref MapInitEvent args)
@@ -90,7 +42,7 @@ public abstract class SharedAirDropSystem : EntitySystem
         }
     }
 
-    private void StartAirDrop(Entity<AirDropComponent?> ent, MapCoordinates pos)
+    protected void StartAirDrop(Entity<AirDropComponent?> ent, MapCoordinates pos)
     {
         if (!Resolve(ent, ref ent.Comp))
             return;
@@ -149,71 +101,5 @@ public abstract class SharedAirDropSystem : EntitySystem
             });
 
         args.Handled = true;
-    }
-
-    private async void OnAirDropSpawn(EntityUid uid, AirDropComponent comp, AirDropSpawnEvent args)
-    {
-        if (args.Handled)
-            return;
-
-        if (!_net.IsServer)
-            return;
-
-        var pos = args.Pos;
-        var supply = Spawn(comp.SupplyDropProto, pos, comp.SupplyDrop);
-        _transform.AttachToGridOrMap(supply);
-        if (Prototype(uid) is { } proto)
-        {
-            EnsureComp<AirDropVisualizerComponent>(supply).SupplyDrop = proto.ID;
-            if (TryComp<AirDropGhostRoleComponent>(supply, out var dropGhost))
-            {
-                dropGhost.SupplyDrop = proto.ID;
-            }
-        }
-
-        QueueDel(uid);
-        args.Handled = true;
-
-        if (comp.SupplyDropTable is { } lootTable)
-        {
-            var forceOpenSupplyDrop = comp.ForceOpenSupplyDrop;
-
-            // пытаемся не сделать smash loot
-            await Timer.Delay(TimeSpan.FromSeconds(1));
-            QueueLocalEvent(new AirDropItemSpawnEvent
-            {
-                DropTable = _prototypeManager.Index(lootTable).Table,
-                SupplyPod = supply,
-                ForceOpenSupplyDrop = forceOpenSupplyDrop
-            });
-        }
-    }
-
-    private void OnAfterInteract(Entity<AirDropItemComponent> ent, ref AfterInteractEvent args)
-    {
-        if (!args.CanReach)
-            return;
-
-        if (!TryComp(ent, out UseDelayComponent? useDelay))
-            return;
-
-        if (_delay.IsDelayed((ent, useDelay)))
-            return;
-
-        _delay.TryResetDelay((ent, useDelay));
-        Spawn(ent.Comp.AirDropProto, Transform(ent).Coordinates);
-    }
-
-    private void OnMapInit(Entity<AirDropItemComponent> ent, ref MapInitEvent args)
-    {
-        if (!TryComp(ent, out UseDelayComponent? useDelay))
-            return;
-        Entity<UseDelayComponent?> dItem = (ent, useDelay);
-        _delay.SetLength(dItem, ent.Comp.Cooldown);
-
-        if (ent.Comp.StartCooldown)
-        {
-            _delay.TryResetDelay((ent, useDelay));
-        }
     }
 }
