+++ Resources/Prototypes/_Backmen/Maps/Pool/backmenstandart.yml
@@ -19,5 +19,4 @@
   - Oasis
   - Packed
   - Plasma
-  - Reach
   - Exo
